pyproject.toml
[project]
name = "mira"
version = "0.0.1"
description = "MIRA — Medical Imaging Recognition Assistant (research prototype, not a medical device)"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

# NOTE: torch / torchvision are NOT listed here. They need CUDA-specific wheels
# (cu128 on the dev box, CPU-only on prod) and are installed out-of-band — see
# backend/ml/README.md. pyproject.toml stays provider-neutral.
dependencies = [
    "monai>=1.3",
    "pydicom>=2.4",
    "SimpleITK>=2.3",
    "nibabel>=5.2",
    "scikit-image>=0.22",
    "scikit-learn>=1.3",
    "numpy>=1.26",
    "pandas>=2.1",
    "pillow>=10.0",
    "matplotlib>=3.8",
    "tqdm>=4.66",
    "kaggle>=1.6",
    "pyyaml>=6.0",
    "structlog>=24.0",
    # Phase 3: REST API
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlmodel>=0.0.19",
    "alembic>=1.13",
    "celery[redis]>=5.4",
    "redis>=5.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.2",
    "minio>=7.2",
    "psycopg2-binary>=2.9",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6",
    "black>=24.8",
    "mypy>=1.10",
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-anyio>=0.0.0",
    "httpx>=0.27",
    "pre-commit>=3.7",
    "ipykernel>=6.29",
    "jupyterlab>=4.2",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["backend"]
include = ["app*", "ml*"]

[tool.ruff]
line-length = 100
target-version = "py311"
# Jupyter notebooks use their own coding style (semicolons to suppress return
# prints, chained one-liners, etc.). Lint them with `nbqa` if you want, but
# don't block commits on normal notebook idioms.
extend-exclude = ["*.ipynb"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
pythonpath = ["backend"]
addopts = "-ra --strict-markers"
